+++ app/src/main/java/com/keycome/twinkleschedule/model/sketch/CourseField.kt
@@ -3,8 +3,8 @@ package com.keycome.twinkleschedule.model.sketch
 import com.keycome.twinkleschedule.model.horizon.Day
 
 sealed class CourseField {
-    data class CourseId(val courseId: Int) : CourseField()
-    data class ParentScheduleId(val parentScheduleId: Int) : CourseField()
+    data class CourseId(val courseId: Long) : CourseField()
+    data class ParentScheduleId(val parentScheduleId: Long) : CourseField()
     data class Title(val title: String) : CourseField()
     data class FDay(val day: Day) : CourseField()
     data class Section(val section: List<Int>) : CourseField()
